pyproject.toml
[tool.poetry]
name = "cats"
version = "0.1.0"
description = ""
authors = ["Pavel <pashapasha-2013@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
